+++ geopolars/src/ops/mod.rs
@@ -1,5 +1,16 @@
 pub mod affine;
-pub mod exterior;
-pub mod length;
+pub(crate) mod area;
+pub(crate) mod centroid;
+pub(crate) mod convex_hull;
+pub(crate) mod distance;
+pub(crate) mod envelope;
+pub(crate) mod explode;
+pub(crate) mod exterior;
+pub(crate) mod geom_type;
+pub(crate) mod is_empty;
+pub(crate) mod is_ring;
+pub(crate) mod length;
+pub(crate) mod point;
 #[cfg(feature = "proj")]
-pub mod proj;
+pub(crate) mod proj;
+pub(crate) mod simplify;
